# src/main/resources/openapi.yaml
openapi: "3.1.0"
info:
  title: "backend_flashpomo API"
  description: "backend_flashpomo API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
paths:
...
  /ensemble/{ensembleId}:
    put:
      summary: "PUT ensemble/{ensembleId}"
      operationId: "updateEnsemble"
      parameters:
        - name: "ensembleId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnsembleRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    delete:
      summary: "DELETE ensemble/{ensembleId}"
      operationId: "deleteEnsemble"
      parameters:
        - name: "ensembleId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    get:
      summary: "GET ensemble/{ensembleId}"
      operationId: "getEnsembleById"
      parameters:
        - name: "ensembleId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
...
  /flashcard/{flashcardId}:
    put:
      summary: "PUT flashcard/{flashcardId}"
      operationId: "updateFlashcard"
      parameters:
        - name: "flashcardId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardRequestEditDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    delete:
      summary: "DELETE flashcard/{flashcardId}"
      operationId: "deleteFlashcard"
      parameters:
        - name: "flashcardId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    get:
      summary: "GET flashcard/{flashcardId}"
      operationId: "getFlashcardById"
      parameters:
        - name: "flashcardId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
  /flashcard/:
    post:
      summary: "POST flashcard/"
      operationId: "createFlashcard"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    get:
      summary: "GET flashcard/"
      operationId: "getAllFlashcards"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Flashcard"
  /flashcard/ensemble/{ensemble}:
    get:
      summary: "GET flashcard/ensemble/{ensemble}"
      operationId: "getEnsembleById"
      parameters:
        - name: "ensemble"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Flashcard"
...
  /pomodoro-story/{pomodoroStoryId}:
    put:
      summary: "PUT pomodoro-story/{pomodoroStoryId}"
      operationId: "updatePomodoroStory"
      parameters:
        - name: "pomodoroStoryId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PomodoroStoryRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    post:
      summary: "POST pomodoro-story/{pomodoroStoryId}"
      operationId: "createPomodoroStory"
      parameters:
        - name: "pomodoroStoryId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PomodoroStoryRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    get:
      summary: "GET pomodoro-story/{pomodoroStoryId}"
      operationId: "getPomodoroStoryById"
      parameters:
        - name: "pomodoroStoryId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
...
  /purchase/{purchaseId}:
    put:
      summary: "PUT purchase/{purchaseId}"
      operationId: "updatePurchase"
      parameters:
        - name: "purchaseId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    post:
      summary: "POST purchase/{purchaseId}"
      operationId: "createPurchase"
      parameters:
        - name: "purchaseId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    delete:
      summary: "DELETE purchase/{purchaseId}"
      operationId: "deletePurchase"
      parameters:
        - name: "purchaseId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    get:
      summary: "GET purchase/{purchaseId}"
      operationId: "getPurchaseById"
      parameters:
        - name: "purchaseId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
...
components:
  schemas:
...
    HttpHeaders:
      type: "object"
      properties: { }
    ResponseEntity:
      type: "object"
      properties:
        headers:
          $ref: "#/components/schemas/HttpHeaders"
      required:
        - "headers"
...
    EnsembleRequestDTO:
      type: "object"
      properties:
        name:
          type: "string"
        description:
          type: "string"
        userId:
          type: "integer"
          format: "int32"
    User:
      type: "object"
      properties:
        userId:
          type: "integer"
          format: "int32"
        name:
          type: "string"
        password:
          type: "string"
        email:
          type: "string"
        role:
          type: "string"
        avatar:
          type: "array"
          items:
            type: "string"
            format: "byte"
        createdAt:
          type: "string"
        updatedAt:
          type: "string"
        status:
          type: "boolean"
    Ensemble:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        name:
          type: "string"
        description:
          type: "string"
        cover:
          type: "array"
          items:
            type: "string"
            format: "byte"
        user:
          $ref: "#/components/schemas/User"
        status:
          type: "boolean"
    FlashcardRequestEditDTO:
      type: "object"
      properties:
        title:
          type: "string"
        content:
          type: "string"
        createdAt:
          type: "string"
        updatedAt:
          type: "string"
        ensembleId:
          type: "integer"
          format: "int32"
        status:
          type: "boolean"
    FlashcardRequestDTO:
      type: "object"
      properties:
        title:
          type: "string"
        content:
          type: "string"
        createdAt:
          type: "string"
        updatedAt:
          type: "string"
        ensembleId:
          type: "integer"
          format: "int32"
    Flashcard:
      type: "object"
      properties:
        flashcardId:
          type: "integer"
          format: "int32"
        title:
          type: "string"
        content:
          type: "string"
        createdAt:
          type: "string"
        updatedAt:
          type: "string"
        ensemble:
          $ref: "#/components/schemas/Ensemble"
        status:
          type: "boolean"
...
    PomodoroStoryRequestDTO:
      type: "object"
      properties:
        pomodoroStoryId:
          type: "integer"
          format: "int64"
        startDateTimestamp:
          type: "string"
        EndDateTimestamp:
          type: "string"
        userId:
          type: "integer"
          format: "int32"
    PurchaseRequestDTO:
      type: "object"
      properties:
        purchaseId:
          type: "integer"
          format: "int64"
        email:
          type: "string"
        rule:
          type: "string"
        avatar:
          type: "array"
          items:
            type: "string"
            format: "byte"
        purschaseDate:
          type: "string"
        status:
          type: "string"
        amount:
          type: "integer"
          format: "int32"
        userId:
          type: "integer"
          format: "int32"